+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/combat/KillAura.kt
@@ -779,6 +779,7 @@ object KillAura : Module("KillAura", Category.COMBAT, Keyboard.KEY_R) {
 
             if (switchMode && !isLookingOnEntities(entity, maxSwitchFOV.toDouble())) continue
 
+            // Credits to Gugustus / Augustus b2.6
             val optimal = (distance * 2.0) + entity.health.toDouble() + (entity.hurtTime.toDouble() * 4.0) + entity.totalArmorValue.toDouble() + (entityFov.toDouble() / 2.0)
 
             val currentValue = when (priority) {